requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
flask-restx==1.1.0
langchain==0.1.3
#pypdf==3.16.4
urllib3==1.26.6
faiss-cpu==1.7.4
python-dotenv==0.17.1